Add the QMX codec as a block codec option
This adds QMX as a potential block codec to ds2i. QMX is a SIMD based
simple16 codec. Some of the tricks in QMX can't be used here as we are
only using it to compress small blocks. Still it looks like it is quite
fast in practice.

All tests pass on my machine.